+++ src/test/java/com/global/api/tests/portico/PorticoEcommerceTests.java
@@ -151,22 +151,24 @@ public class PorticoEcommerceTests {
         Transaction response = card.charge(new BigDecimal("11"))
                 .withCurrency("USD")
                 .withCustomerData(customer)
+                .withAllowDuplicates(true)
                 .execute();
 
         assertNotNull(response);
         assertEquals("00", response.getResponseCode());
 
         TransactionSummary txnSummary = ReportingService.transactionDetail(response.getTransactionId()).execute();
-
         assertNotNull(txnSummary);
         assertEquals(customerEmail, txnSummary.getEmail() );
     }
     @Test
     public void testCardHolderEmailIfCustomerDataIsNull() throws ApiException {
         Customer customerData = null;
+
         Transaction response = card.charge(new BigDecimal("11"))
                 .withCurrency("USD")
                 .withCustomerData(customerData)
+                .withAllowDuplicates(true)
                 .execute();
 
         assertNotNull(response);