+++ src/test/java/com/global/api/tests/terminals/upa/UpaCreditTests.java
@@ -31,13 +31,11 @@ public class UpaCreditTests {
     public UpaCreditTests() throws ApiException {
         ConnectionConfig config = new ConnectionConfig();
         config.setPort(8081);
-        config.setIpAddress("10.253.146.155");
+        config.setIpAddress("192.168.0.199");
         config.setTimeout(45000);
         config.setRequestIdProvider(new RandomIdProvider());
         config.setDeviceType(DeviceType.UPA_DEVICE);
         config.setConnectionMode(ConnectionModes.TCP_IP);
-
-//        config.setRequestLogger(new RequestFileLogger("creditTests.txt"));
 
         device = DeviceService.create(config);
         assertNotNull(device);